+++ Gemfile
@@ -7,7 +7,6 @@ ruby "2.6.5"
 gem "dotenv"
 gem "httparty"
 gem "rake"
-gem "require_all"
 gem "rollbar"
 gem "slack-ruby-client"
 gem "snitcher"
