+++ Rakefile
@@ -15,7 +15,7 @@ namespace :reports do
 
   desc "This task send report for the past day"
   task :daily do
-    HarvestNotifier::Base.new.create_daily_report
+    HarvestNotifier.create_daily_report
     Snitcher.snitch(ENV["SNITCH_DAILY"]) if ENV["SNITCH_DAILY"]
   end
 end
